Created README.md
Hi, added some presentation-style to your project with this README file. Just gives a description of what the project is about for anyone looking. Idk if you are looking to pursue software development particularly; if you are then having README files on your projects helps viewers see what kind of work you've done at a glance.

You don't have to merge this with your project if you don't want to, and then no hard feelings of course :) I'm just looking for ways to contribute to open source and figured it'd be nice to help out someone I've worked with before.